backend/WhereMoney.Api/pyproject.toml
[tool.poetry]
name = "pythonproject"
version = "0.1.0"
description = ""
authors = ["DavideRodo <you@example.com>"]

[tool.poetry.dependencies]
python = ">=3.12"
flask = "3.1.2"
flask-restx = "1.3.0"
flask-migrate = "4.1.0"
flask-cors = "5.0.0"
flask-sqlalchemy = "3.1.1"
pyjwt = "2.10.1"
bcrypt = "4.3.0"
alembic = "1.16.5"
sqlalchemy = "2.0.43"
werkzeug = "3.1.3"
jinja2 = "3.1.6"
pandas = "^2.3.2"
psycopg2-binary = "^2.9.7"

[build-system]
requires = ["poetry-core>=1.6.0"]
build-backend = "poetry.core.masonry.api"
